+++ frontend/packages/app/src/pages/project-details/context.ts
@@ -15,36 +15,6 @@ export type RepositoryInput = Pick<
 > &
   Partial<Pick<ProjectRepositoryConnection, "name">>;
 
-export type CreateRateInput = {
-  employee?: string;
-  hourlyRate: number;
-  validFrom: string;
-};
-
-export type EditRateInput = {
-  name: string;
-  employee: string;
-  hourlyRate: number;
-  validFrom: string;
-};
-
-export type CreateContractInput = {
-  startDate: string;
-  endDate: string;
-  hoursBought: number;
-  salesOrder?: string;
-  salesInvoice?: string;
-};
-
-export type EditContractInput = {
-  name: string;
-  startDate: string;
-  endDate: string;
-  hoursBought: number;
-  salesOrder?: string;
-  salesInvoice?: string;
-};
-
 export interface ProjectDetailContextProps {
   projectId: string;
   project: FrappeDoc<ProjectDoc> | undefined;
@@ -55,12 +25,6 @@ export interface ProjectDetailContextProps {
   addMember: (userId: string) => Promise<void>;
   removeMember: (userId: string) => Promise<void>;
   updateContacts: (contactIds: string[]) => Promise<void>;
-  deleteRate: (name: string) => Promise<void>;
-  createRate: (input: CreateRateInput) => Promise<void>;
-  editRate: (input: EditRateInput) => Promise<void>;
-  createContract: (input: CreateContractInput) => Promise<void>;
-  editContract: (input: EditContractInput) => Promise<void>;
-  deleteContract: (name: string) => Promise<void>;
 }
 
 const noop = () => {};
@@ -76,12 +40,6 @@ export const ProjectDetailContext = createContext<ProjectDetailContextProps>({
   addMember: asyncNoop,
   removeMember: asyncNoop,
   updateContacts: asyncNoop,
-  deleteRate: asyncNoop,
-  createRate: asyncNoop,
-  editRate: asyncNoop,
-  createContract: asyncNoop,
-  editContract: asyncNoop,
-  deleteContract: asyncNoop,
 });
 
 export const useProjectDetail = <T>(
