+++ frontend/packages/app/src/pages/project-details/tabs/tracking/components/contractModal/types.ts
@@ -1,4 +1,4 @@
-import type { CreateContractInput } from "../../../../context";
+import type { CreateContractInput } from "../../context";
 
 export type ContractModalMode = "add" | "edit";
