+++ frontend/packages/app/src/pages/project-details/tabs/tracking/context.ts
@@ -6,75 +6,15 @@ import { createContext, useContextSelector } from "use-context-selector";
 /**
  * Internal dependencies.
  */
-import type { ContractRow, RateRow } from "./types";
 import type {
-  CreateContractInput,
+  ContractRow,
+  RateRow,
+  Tracking,
   CreateRateInput,
-  EditContractInput,
   EditRateInput,
-} from "../../context";
-
-export type ProjectFlatRate = {
-  flat_rate_hourly?: number;
-  flat_rate_valid_from?: string;
-};
-
-export type ProjectRate = {
-  name?: string | null;
-  employee?: string;
-  employee_name?: string;
-  hourly_billing_rate?: number;
-  valid_from?: string;
-};
-
-export type TrackingTasks = {
-  total: number;
-  open: number;
-  completed: number;
-};
-
-export type TrackingInvoiceBurn = {
-  currency: string;
-  invoiced_and_paid: number;
-  invoiced_but_not_paid: number;
-  total_project_amount: number;
-};
-
-export type TrackingContract = {
-  name: string;
-  start_date: string;
-  end_date: string;
-  hours_purchased: number;
-  consumed_hours: number;
-  remaining_hours: number;
-  sales_order: string;
-  sales_invoice: string;
-};
-
-export type TrackingMessage = {
-  company: string;
-  billing_type: string;
-  currency: string;
-  total_project_value: number;
-  project_profit: number;
-  projected_profit_margin: number;
-  actual_cost_incurred: number;
-  forecasted_cost_to_completion: number;
-  expected_total_cost: number;
-  hours_utilised: number;
-  hours_remaining: number | null;
-  tasks: TrackingTasks;
-  invoice_burn: TrackingInvoiceBurn;
-  contracts: TrackingContract[] | null;
-  project_rates: [ProjectFlatRate, ...ProjectRate[]] | null;
-  lifetime_value_to_date: number;
-  expected_lifetime_value: number;
-  lifetime_value_vs_billed_amount: number;
-};
-
-export type Response = { message: TrackingMessage };
-
-export type Tracking = TrackingMessage;
+  CreateContractInput,
+  EditContractInput,
+} from "./types";
 
 export interface TrackingContextProps {
   tracking: Tracking;
