+++ frontend/packages/app/src/pages/project-details/tabs/tracking/provider.tsx
@@ -2,36 +2,43 @@
  * External dependencies.
  */
 import { useCallback, useMemo, useState, type PropsWithChildren } from "react";
-import { useFrappeGetCall } from "frappe-react-sdk";
+import { useToasts } from "@rtcamp/frappe-ui-react";
+import {
+  type FrappeError,
+  useFrappeGetCall,
+  useFrappeUpdateDoc,
+} from "frappe-react-sdk";
 
 /**
  * Internal dependencies.
  */
-import { currencyFormat } from "@/lib/utils";
+import { currencyFormat, parseFrappeErrorMsg } from "@/lib/utils";
 import {
   DEFAULT_TRACKING,
   TrackingContext,
-  type Response,
-  type Tracking,
   type TrackingContextProps,
 } from "./context";
-import type { ContractRow, RateRow } from "./types";
-import {
-  useProjectDetail,
-  type CreateContractInput,
-  type CreateRateInput,
-  type EditContractInput,
-  type EditRateInput,
-} from "../../context";
+
+import type {
+  ContractRow,
+  RateRow,
+  CreateContractInput,
+  CreateRateInput,
+  EditContractInput,
+  EditRateInput,
+  Response,
+  Tracking,
+} from "./types";
+
+import { useProjectDetail } from "../../context";
 
 export function TrackingProvider({ children }: PropsWithChildren) {
   const projectId = useProjectDetail((s) => s.projectId);
-  const deletedRateDoc = useProjectDetail((s) => s.deleteRate);
-  const createRateDoc = useProjectDetail((s) => s.createRate);
-  const editRateDoc = useProjectDetail((s) => s.editRate);
-  const createContractDoc = useProjectDetail((s) => s.createContract);
-  const editContractDoc = useProjectDetail((s) => s.editContract);
-  const deleteContractDoc = useProjectDetail((s) => s.deleteContract);
+  const project = useProjectDetail((s) => s.project);
+  const projectMutate = useProjectDetail((s) => s.mutate);
+  const { updateDoc } = useFrappeUpdateDoc();
+  const toast = useToasts();
+
   const [addRateModalOpen, setAddRateModalOpen] = useState(false);
   const [editingRate, setEditingRate] = useState<RateRow | null>(null);
   const [addContractModalOpen, setAddContractModalOpen] = useState(false);
@@ -53,50 +60,152 @@ export function TrackingProvider({ children }: PropsWithChildren) {
 
   const deleteRate = useCallback(
     async (name: string) => {
-      await deletedRateDoc(name);
-      await mutate();
+      const current = project?.custom_project_billing_team ?? [];
+      try {
+        await updateDoc("Project", projectId, {
+          custom_project_billing_team: current.filter(
+            (row) => row.name !== name,
+          ),
+        });
+        projectMutate();
+        await mutate();
+        toast.success("Rate deleted");
+      } catch (err) {
+        toast.error(parseFrappeErrorMsg(err as FrappeError));
+      }
     },
-    [deletedRateDoc, mutate],
-  );
-
-  const createRate = useCallback(
-    async (input: CreateRateInput) => {
-      await createRateDoc(input);
-      await mutate();
-    },
-    [createRateDoc, mutate],
+    [updateDoc, projectId, projectMutate, mutate, project, toast],
   );
 
   const editRate = useCallback(
-    async (input: EditRateInput) => {
-      await editRateDoc(input);
+    async ({ name, employee, hourlyRate, validFrom }: EditRateInput) => {
+      const current = project?.custom_project_billing_team ?? [];
+      try {
+        await updateDoc("Project", projectId, {
+          custom_project_billing_team: current.map((row) =>
+            row.name === name
+              ? {
+                  ...row,
+                  employee,
+                  hourly_billing_rate: hourlyRate,
+                  valid_from: validFrom,
+                }
+              : row,
+          ),
+        });
+        projectMutate();
+        await mutate();
+        toast.success("Rate updated");
+      } catch (err) {
+        toast.error(parseFrappeErrorMsg(err as FrappeError));
+      }
+    },
+    [updateDoc, projectId, projectMutate, mutate, project, toast],
+  );
+
+  const createRate = useCallback(
+    async ({ employee, hourlyRate, validFrom }: CreateRateInput) => {
+      const current = project?.custom_project_billing_team ?? [];
+      await updateDoc("Project", projectId, {
+        custom_project_billing_team: [
+          ...current,
+          {
+            employee: employee ?? "",
+            hourly_billing_rate: hourlyRate,
+            valid_from: validFrom,
+          },
+        ],
+      });
+      projectMutate();
       await mutate();
     },
-    [editRateDoc, mutate],
+    [updateDoc, projectId, projectMutate, mutate, project],
   );
 
   const createContract = useCallback(
-    async (input: CreateContractInput) => {
-      await createContractDoc(input);
-      await mutate();
+    async ({
+      startDate,
+      endDate,
+      hoursBought,
+      salesOrder,
+      salesInvoice,
+    }: CreateContractInput) => {
+      const current = project?.custom_project_budget_hours ?? [];
+      try {
+        await updateDoc("Project", projectId, {
+          custom_project_budget_hours: [
+            ...current,
+            {
+              start_date: startDate,
+              end_date: endDate,
+              hours_purchased: hoursBought,
+              sales_order: salesOrder ?? "",
+              sales_invoice: salesInvoice ?? "",
+            },
+          ],
+        });
+        projectMutate();
+        await mutate();
+        toast.success("Contract added");
+      } catch (err) {
+        toast.error(parseFrappeErrorMsg(err as FrappeError));
+      }
     },
-    [createContractDoc, mutate],
+    [updateDoc, projectId, projectMutate, mutate, project, toast],
   );
 
   const editContract = useCallback(
-    async (input: EditContractInput) => {
-      await editContractDoc(input);
-      await mutate();
+    async ({
+      name,
+      startDate,
+      endDate,
+      hoursBought,
+      salesOrder,
+      salesInvoice,
+    }: EditContractInput) => {
+      const current = project?.custom_project_budget_hours ?? [];
+      try {
+        await updateDoc("Project", projectId, {
+          custom_project_budget_hours: current.map((row) =>
+            row.name === name
+              ? {
+                  ...row,
+                  start_date: startDate,
+                  end_date: endDate,
+                  hours_purchased: hoursBought,
+                  sales_order: salesOrder ?? "",
+                  sales_invoice: salesInvoice ?? "",
+                }
+              : row,
+          ),
+        });
+        projectMutate();
+        await mutate();
+        toast.success("Contract updated");
+      } catch (err) {
+        toast.error(parseFrappeErrorMsg(err as FrappeError));
+      }
     },
-    [editContractDoc, mutate],
+    [updateDoc, projectId, projectMutate, mutate, project, toast],
   );
 
   const deleteContract = useCallback(
     async (name: string) => {
-      await deleteContractDoc(name);
-      await mutate();
+      const current = project?.custom_project_budget_hours ?? [];
+      try {
+        await updateDoc("Project", projectId, {
+          custom_project_budget_hours: current.filter(
+            (row) => row.name !== name,
+          ),
+        });
+        projectMutate();
+        await mutate();
+        toast.success("Contract deleted");
+      } catch (err) {
+        toast.error(parseFrappeErrorMsg(err as FrappeError));
+      }
     },
-    [deleteContractDoc, mutate],
+    [updateDoc, projectId, projectMutate, mutate, project, toast],
   );
 
   const value = useMemo<TrackingContextProps>(() => {
